+++ front/constants/home.ts
@@ -10,4 +10,4 @@ export const totalTrainingTime = createSelectOptions([ 3, 5, 10 ]);
 export const trainingTime = createSelectOptions([ 30, 60 ]);
 export const breakTime = createSelectOptions([ 10, 20 ]);
 
-export const trainingInfoItems = ['categoryId', 'totalTrainingTime', 'trainingTime', 'breakTime'];
+export const trainingInfoItems = ['totalTrainingTime', 'trainingTime', 'breakTime'];
